move bin/dswx_hls functions to modules/core and modules/dswx_hls
